+++ packages/config/tailwind-config/tailwind.config.js
@@ -12,6 +12,8 @@ const path = require('path');
  */
 const NEXTUI_THEME_USED_COMPONENT_PATHS = [
   'autocomplete',
+  'avatar',
+  'accordion',
   'button',
   'card',
   'chip',