+++ packages/pixels/src/Accordion/index.ts
@@ -1,0 +1,7 @@
+import Accordion from './Accordion';
+
+export type { AccordionProps } from './Accordion';
+
+export { Accordion };
+
+export default Accordion;