+++ packages/pixels/src/index.ts
@@ -1,4 +1,7 @@
 // components
+export * from './Accordion';
+export * from './Avatar';
+export * from './AvatarGroup';
 export * from './Button';
 export * from './Card';
 export * from './Json';