+++ packages/veto/src/types/discriminatedUnion/discriminatedUnion.ts
@@ -1,3 +1,17 @@
+import type { ZodDiscriminatedUnionOption } from 'zod';
+
 import { z } from 'zod';
 
-export default z.discriminatedUnion;
+// eslint-disable-next-line prefer-destructuring
+export const discriminatedUnion: <
+  Discriminator extends string,
+  Types extends [
+    ZodDiscriminatedUnionOption<Discriminator>,
+    ...ZodDiscriminatedUnionOption<Discriminator>[],
+  ],
+>(
+  discriminator: Discriminator,
+  options: Types,
+) => z.ZodDiscriminatedUnion<Discriminator, Types> = z.discriminatedUnion;
+
+export type VDiscriminatedUnion = typeof discriminatedUnion;