+++ packages/veto/src/types/literal/literal.ts
@@ -1,3 +1,6 @@
 import { z } from 'zod';
 
-export default z.literal;
+// eslint-disable-next-line prefer-destructuring
+export const literal = z.literal;
+
+export type VLiteral = typeof literal;