+++ packages/veto/src/types/number/number.ts
@@ -1,7 +1,12 @@
-import type { ZodNumber } from 'zod';
-
 import { z } from 'zod';
 
-export type VNumber = ZodNumber;
+// eslint-disable-next-line prefer-destructuring
+export const number = z.number;
 
-export default z.number;
+export type VNumber = typeof number;
+export type VVNumberSchema = ReturnType<VNumber>;
+
+/** when used with refine or superRefine */
+export type VNumberRefined<Options = undefined> = (
+  options?: Options,
+) => z.ZodEffects<VVNumberSchema, string, string>;