+++ packages/veto/src/types/record/record.ts
@@ -1,3 +1,6 @@
 import { z } from 'zod';
 
-export default z.record;
+// eslint-disable-next-line prefer-destructuring
+export const record = z.record;
+
+export type VRecord = typeof record;