+++ packages/veto/src/types/vEnum/vEnum.ts
@@ -1,3 +1,9 @@
+import type { ZodEnum } from 'zod';
+
 import { z } from 'zod';
 
-export default z.enum;
+// eslint-disable-next-line prefer-destructuring
+export const vEnum = z.enum;
+
+export type VEnum = typeof vEnum;
+export type VEnumSchema<T extends [string, ...string[]]> = ZodEnum<T>;