+++ packages/veto/CHANGELOG.md
@@ -1,5 +1,241 @@
 # Changelog
 
+## [0.4.0](https://github.com/fuf-stack/uniform/compare/veto-v0.3.32...veto-v0.4.0) (2024-05-27)
+
+
+### Features
+
+* **veto:** add objectLoose ([bed4cff](https://github.com/fuf-stack/uniform/commit/bed4cffecf78c165fbc69ed436d35032a8e4a189))
+
+## [0.3.32](https://github.com/fuf-stack/uniform/compare/veto-v0.3.31...veto-v0.3.32) (2024-05-23)
+
+
+### Bug Fixes
+
+* **veto:** refactor exports ([e079b27](https://github.com/fuf-stack/uniform/commit/e079b2781804297624e44aa4ea2ec5a403d5a31a))
+
+## [0.3.31](https://github.com/fuf-stack/uniform/compare/veto-v0.3.30...veto-v0.3.31) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** add type anotation to discriminatedUnion ([2ccd2ee](https://github.com/fuf-stack/uniform/commit/2ccd2eeffa303c94b7e13b291b13d19cc35e7a8d))
+
+## [0.3.30](https://github.com/fuf-stack/uniform/compare/veto-v0.3.29...veto-v0.3.30) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** remove zod from peerDeps ([45bd8e9](https://github.com/fuf-stack/uniform/commit/45bd8e9a186845a9ed6cbce10454560905d56f4e))
+
+## [0.3.29](https://github.com/fuf-stack/uniform/compare/veto-v0.3.28...veto-v0.3.29) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** bundle zod types again with --dts-resolve ([2e4e698](https://github.com/fuf-stack/uniform/commit/2e4e6980821fe6a547b9581a9d1ad667800add7d))
+
+## [0.3.28](https://github.com/fuf-stack/uniform/compare/veto-v0.3.27...veto-v0.3.28) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** refactor some portable type anotations ([0569b95](https://github.com/fuf-stack/uniform/commit/0569b952dadf3de74637d749976c0f2bb1fce6b5))
+
+## [0.3.27](https://github.com/fuf-stack/uniform/compare/veto-v0.3.26...veto-v0.3.27) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** add VetoNullable, VArraySchema and VArraySchema ([6d42799](https://github.com/fuf-stack/uniform/commit/6d427995a67eeef5857084b83bb295e5b2f8d480))
+
+## [0.3.26](https://github.com/fuf-stack/uniform/compare/veto-v0.3.25...veto-v0.3.26) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** add VObjectRefined type ([874dd00](https://github.com/fuf-stack/uniform/commit/874dd0070aa5714bbcf1823e065ec9a9f374b7b9))
+* **veto:** use VetoEffects  in VObjectRefined ([dd284d5](https://github.com/fuf-stack/uniform/commit/dd284d5b757d1673accc09dd40d41f9164cf1a8b))
+
+## [0.3.25](https://github.com/fuf-stack/uniform/compare/veto-v0.3.24...veto-v0.3.25) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** add type VEnumSchema ([3cd9156](https://github.com/fuf-stack/uniform/commit/3cd915675643eeaaf233bc806f400af3a400551b))
+
+## [0.3.24](https://github.com/fuf-stack/uniform/compare/veto-v0.3.23...veto-v0.3.24) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** add VObjectSchema type ([4c4aa8a](https://github.com/fuf-stack/uniform/commit/4c4aa8a5ed13bfe6c6978cfbe4feb32fb8489ad6))
+
+## [0.3.23](https://github.com/fuf-stack/uniform/compare/veto-v0.3.22...veto-v0.3.23) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** add type VetoOptional ([486a7a1](https://github.com/fuf-stack/uniform/commit/486a7a1908528f02015ba03bdb48b3ff6c42874e))
+
+## [0.3.22](https://github.com/fuf-stack/uniform/compare/veto-v0.3.21...veto-v0.3.22) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** add options to VStringRefined and add VNumberRefined ([3fc48c7](https://github.com/fuf-stack/uniform/commit/3fc48c798da032cac3db6f37fe8c652efd82e79d))
+
+## [0.3.21](https://github.com/fuf-stack/uniform/compare/veto-v0.3.20...veto-v0.3.21) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** add VStringRefined type ([f7d3dd2](https://github.com/fuf-stack/uniform/commit/f7d3dd2c0aeb2e96d82eee5adee2cfd37f122a24))
+
+## [0.3.20](https://github.com/fuf-stack/uniform/compare/veto-v0.3.19...veto-v0.3.20) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** use direct string return type ([2803b0d](https://github.com/fuf-stack/uniform/commit/2803b0d91e2f70011c3d684afaac76dffcc24bd1))
+
+## [0.3.19](https://github.com/fuf-stack/uniform/compare/veto-v0.3.18...veto-v0.3.19) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** create dts with tsup ([cabc23d](https://github.com/fuf-stack/uniform/commit/cabc23d4b0347706add79137f169f8c8d7dda8ec))
+
+## [0.3.18](https://github.com/fuf-stack/uniform/compare/veto-v0.3.17...veto-v0.3.18) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** try to make string portable ([2a01afa](https://github.com/fuf-stack/uniform/commit/2a01afad1ba3c93c59e84aaeda5ad3b41dc1fc6c))
+
+## [0.3.17](https://github.com/fuf-stack/uniform/compare/veto-v0.3.16...veto-v0.3.17) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** refactor some schema types ([4805648](https://github.com/fuf-stack/uniform/commit/480564898eb1d2858ff98b798c62ab333dcb5676))
+* **veto:** refactor some types ([9dd9380](https://github.com/fuf-stack/uniform/commit/9dd938096bc87b22f14fa1ef9b4733d21d23cced))
+
+## [0.3.16](https://github.com/fuf-stack/uniform/compare/veto-v0.3.15...veto-v0.3.16) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** do not inline zod types ([6c313b1](https://github.com/fuf-stack/uniform/commit/6c313b1399db76280273203f1e02377a37d70b11))
+
+## [0.3.15](https://github.com/fuf-stack/uniform/compare/veto-v0.3.14...veto-v0.3.15) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** use dts-bundle-generator to generate types bundle ([b82220b](https://github.com/fuf-stack/uniform/commit/b82220b76f8a97746e7b02f0e115482b940415b7))
+
+## [0.3.14](https://github.com/fuf-stack/uniform/compare/veto-v0.3.13...veto-v0.3.14) (2024-05-22)
+
+
+### Bug Fixes
+
+* **veto:** make all validators portable ([fc9bf0b](https://github.com/fuf-stack/uniform/commit/fc9bf0b11b0acae96e8cdfee54064bb7f4a76222))
+
+## [0.3.13](https://github.com/fuf-stack/uniform/compare/veto-v0.3.12...veto-v0.3.13) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** use --dts-resolve ([c3e8781](https://github.com/fuf-stack/uniform/commit/c3e8781164045e294b070741b1d67a744ef1466b))
+
+## [0.3.12](https://github.com/fuf-stack/uniform/compare/veto-v0.3.11...veto-v0.3.12) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** improve protable type of string ([b161e51](https://github.com/fuf-stack/uniform/commit/b161e51c0403ffd8c28d437aa4752cd3edcb1afd))
+
+## [0.3.11](https://github.com/fuf-stack/uniform/compare/veto-v0.3.10...veto-v0.3.11) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** improve protable type of string ([7ffa34e](https://github.com/fuf-stack/uniform/commit/7ffa34e3af0bc3a11c210578b959841c7e54d29a))
+
+## [0.3.10](https://github.com/fuf-stack/uniform/compare/veto-v0.3.9...veto-v0.3.10) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** improve protable type of string ([516cc5d](https://github.com/fuf-stack/uniform/commit/516cc5dd0de21963156f37ec53691bf389d0330f))
+
+## [0.3.9](https://github.com/fuf-stack/uniform/compare/veto-v0.3.8...veto-v0.3.9) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** add protable type to string ([7c731d5](https://github.com/fuf-stack/uniform/commit/7c731d53a79138123083e54a35163c5a1114b4f4))
+
+## [0.3.8](https://github.com/fuf-stack/uniform/compare/veto-v0.3.7...veto-v0.3.8) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** do not use --dts-resolve ([15e0baf](https://github.com/fuf-stack/uniform/commit/15e0baf4d8b9031284042903e13242e1302fbb9d))
+
+## [0.3.7](https://github.com/fuf-stack/uniform/compare/veto-v0.3.6...veto-v0.3.7) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** add corrext protable type for VObject ([b22f8c1](https://github.com/fuf-stack/uniform/commit/b22f8c1a6b88fed5f07049799388cff5e6fa31fb))
+
+## [0.3.6](https://github.com/fuf-stack/uniform/compare/veto-v0.3.5...veto-v0.3.6) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** cleanup index exports ([6f26f94](https://github.com/fuf-stack/uniform/commit/6f26f94747dc9c5c29653ea73507cc67d52078cd))
+
+## [0.3.5](https://github.com/fuf-stack/uniform/compare/veto-v0.3.4...veto-v0.3.5) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** include zod types with --dts-resolve ([684e9b6](https://github.com/fuf-stack/uniform/commit/684e9b6d6101f4aff7e10b187d95ffaf9e2447ac))
+
+## [0.3.4](https://github.com/fuf-stack/uniform/compare/veto-v0.3.3...veto-v0.3.4) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** try to bundle zod resolved ([f424891](https://github.com/fuf-stack/uniform/commit/f4248919018aa32fee3a7feec6ef649ecf2fb592))
+
+## [0.3.3](https://github.com/fuf-stack/uniform/compare/veto-v0.3.2...veto-v0.3.3) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** bundle zod types with dts-bundle-generator ([82a2c29](https://github.com/fuf-stack/uniform/commit/82a2c29495420a6e3fd7ef0e0312fc752cc94a09))
+* **veto:** try to bundle zod correctly ([b2c61bd](https://github.com/fuf-stack/uniform/commit/b2c61bdd31489844e142be5ed2135980557fa911))
+
+## [0.3.2](https://github.com/fuf-stack/uniform/compare/veto-v0.3.1...veto-v0.3.2) (2024-05-21)
+
+
+### Bug Fixes
+
+* **veto:** try to export zod types to make them portable ([9f28172](https://github.com/fuf-stack/uniform/commit/9f28172fa3cba2f1bec203cb2a2c60039217afb1))
+
+## [0.3.1](https://github.com/fuf-stack/uniform/compare/veto-v0.3.0...veto-v0.3.1) (2024-05-20)
+
+
+### Bug Fixes
+
+* **deps:** update dependency validator to v13.12.0 ([145ad79](https://github.com/fuf-stack/uniform/commit/145ad794377f0cc17c58b735026c831cdb9468bc))
+* **deps:** update dependency zod to v3.23.7 ([ba5b0dd](https://github.com/fuf-stack/uniform/commit/ba5b0ddce7bd616fbd1fbdfb657a90565c56f424))
+* **deps:** update dependency zod to v3.23.8 ([4d8fe8a](https://github.com/fuf-stack/uniform/commit/4d8fe8a678e23912b50bd28561033ce87e099130))
+
 ## [0.3.0](https://github.com/fuf-stack/uniform/compare/veto-v0.2.1...veto-v0.3.0) (2024-05-06)